Install the list of generated META files
If findlib was already installed (either an old version or even just
rebuilding the current version) then configure would detect a
findlib-generated META file as a third-party META file and so skip it.
This meant that for OCaml 4.05 and earlier, the num package META would
not be generated on reinstall or upgrade unless the current findlib was
first removed. opam doesn't do that and the INSTALL instructions don't
state it either.

configure no longer checks for the presence of pre-installed META files.
Instead, the list of META files which will be generated is written to
Makefile.packages and this file is installed as part of the findlib
ocamlfind package (so cat `ocamlfind query findlib`/Makefile.packages
will reveal its content). If this file is found _in the installation_
then configure will read and parse the value from it. This allows
configure to differentiate between a package which was installed by
findlib and a package which has been installed separately.

The complex part is what to do with findlib 1.8.1 and earlier
installations which don't have this file. In this instance, configure
runs its detection as normal. For example, in OCaml 4.06 and later with
the separate num package installed, configure will now cause the num and
num-top META files to be generated and the num-top library to be built.
However, it also records that the information was fuzzy and should be
checked again prior to installation. At make install/uninstall, the
installed META files are probed for their content and installation of
the package only takes place if they contain the '[distributed with
Ocaml]' line.