Interpret relative paths to -sitelib
If -sitelib is passed a relative path, this is interpreted relative to
the path used for findlib.conf (i.e. the path given for -config). For
example, `./configure -config $(opam var lib)/findlib.conf -sitelib .`
has the same meaning as passing `-sitelib $(opam var lib)`.

If Relocatable OCaml is being used, and the OCaml Standard Library is in
a directory underneath `-config` (as an opam switch does, for example),
and the ocamlfind binary is being installed to the same directory as the
compiler, then Relocatable Findlib is enabled. In mode, `ocamlfind`
searches for `findlib.conf` relative to the location of OCaml's Standard
Library and `topfind` similarly locates the `-sitelib` directory
relative to OCaml's Standard Library. `findlib.conf` is also configured
to use relative paths.