Use Unix.realpath when available checking ld.conf
When Findlib is configured to use a shared stublibs directory (which, in
particular, is the configuration opam uses), ocamlfind checks to see
that the shared stublibs directory is mentioned in ld.conf.

Previously, this check tries to take into account path normalisation and
case-insensitive file-systems, but the check is thwarted by OCaml 5.4's
new explicit-relative lines in ld.conf (for opam, it therefore can't
recognise that ../stublibs refers to `$(opam var lib)/stublibs`).

OCaml 4.13+ has Unix.realpath, so we detect the presence of that and use
it to canonicalise the paths. This works reliably because all the
directories are known to exist at the point of the check. Earlier
versions (which may also use the same ld.conf syntax owing to the
Relocatable OCaml backports) will continue to display the innocuous
warning.